modularize account selection
