+++ components/ui/ChooseWalletOverlay.vue
@@ -1,0 +1,118 @@
+<template>
+  <OverlayDialog :show="show" :close="close" title="Access Your Wallet!">
+    <div class="mt-2">
+      <p class="text-sm text-gray-400">How would you like to connect?</p>
+      <div class="mt-4">
+        <button
+          @click="createTestingAccount"
+          class="incognitee-bg btn btn_gradient rounded-md px-3.5 py-2.5 text-sm font-semibold text-white shadow-sm hover:bg-indigo-400 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-indigo-400"
+        >
+          Create a New Account for Testing
+        </button>
+      </div>
+      <p class="mt-4">or</p>
+      <div v-if="extensionAccounts.length < 1" class="mt-4 flex flex-col">
+        <div
+          class="mx-auto grid max-w-lg grid-cols-2 gap-x-3 gap-y-3 sm:max-w-xl sm:grid-cols-4 sm:gap-x-3 lg:mx-0 lg:max-w-none lg:grid-cols-4"
+        >
+          <a href="https://talisman.xyz/download"
+            ><img
+              class="col-span-1 max-h-10 w-full object-contain lg:col-span-1"
+              src="/img/index/talisman-logo.svg"
+              alt="talisman"
+          /></a>
+          <a href="https://novawallet.io/"
+            ><img
+              class="col-span-1 max-h-7 w-full object-contain lg:col-span-1"
+              src="/img/index/nova-wallet-logo.svg"
+              alt="nova wallet"
+          /></a>
+          <a href="https://www.subwallet.app/"
+            ><img
+              class="col-span-1 max-h-10 w-full object-contain lg:col-span-1"
+              src="/img/index/sub-wallet-logo.svg"
+              alt="sub wallet"
+          /></a>
+          <a href="https://polkadot.js.org/extension/"
+            ><img
+              class="col-span-1 max-h-7 w-full object-contain lg:col-span-1"
+              src="/img/index/polkadotjs-logo.svg"
+              alt="polkajs"
+          /></a>
+        </div>
+        <div class="mt-10">
+          <button
+            @click="connectExtension"
+            class="incognitee-bg btn btn_gradient rounded-md px-3.5 py-2.5 text-sm font-semibold text-white shadow-sm hover:bg-indigo-400 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-indigo-400"
+          >
+            Connect Signer Extension
+          </button>
+        </div>
+      </div>
+      <div
+        v-if="extensionAccounts.length > 0"
+        ref="walletSection"
+        id="wallet"
+        class="py-12 sm:py-16"
+      >
+        <p class="text-sm text-gray-400">
+          Choose one of your extension accounts
+        </p>
+        <select
+          v-model="selectedExtensionAccount"
+          id="account.address"
+          name="account.address"
+          placeholder="account.address"
+          class="w-full rounded-md border-0 bg-gray-800 py-1.5 text-white shadow-sm ring-1 ring-inset ring-gray-700 focus:ring-1 focus:ring-inset focus:ring-incognitee-green sm:text-sm sm:leading-6"
+        >
+          <option
+            v-for="account in extensionAccounts"
+            :key="account.address"
+            :value="account.address"
+          >
+            {{ account.meta.name }}
+          </option>
+        </select>
+      </div>
+      <div v-if="accountStore.hasInjector" class="mt-10">
+        <p>
+          please allow this app to read your balance by signing the upcoming
+          request in your extension
+        </p>
+        <p>this window will close once a balance could be fetched</p>
+      </div>
+    </div>
+  </OverlayDialog>
+</template>
+
+<script setup lang="ts">
+import {
+  connectExtension,
+  extensionAccounts,
+} from "@/lib/signerExtensionUtils";
+import OverlayDialog from "@/components/ui/OverlayDialog.vue";
+import { ref } from "vue";
+import { isLive } from "@/lib/environmentConfig";
+import { useAccount } from "@/store/account.ts";
+
+const accountStore = useAccount();
+
+import { defineProps } from "vue";
+
+const props = defineProps({
+  createTestingAccount: {
+    type: Function,
+    required: false,
+  },
+  show: {
+    type: Boolean,
+    required: true,
+  },
+  close: {
+    type: Function,
+    required: false,
+  },
+});
+</script>
+
+<style scoped></style>
